refactor(gateway): renew gateway certificate returns more info about success
